Split up tensor_model.hpp to avoid ODR violation.
