+++ eval/src/vespa/eval/eval/test/CMakeLists.txt
@@ -6,6 +6,7 @@ vespa_add_library(eval_eval_test OBJECT
     reference_evaluation.cpp
     reference_operations.cpp
     tensor_conformance.cpp
+    tensor_model.cpp
     test_io.cpp
     value_compare.cpp
 )
